apps/trainer/app/main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
import uvicorn
import redis
import os
from dotenv import load_dotenv

# 加载环境变量
load_dotenv()

# 创建 FastAPI 应用
app = FastAPI(
    title="XGBoost Training Service",
    description="XGBoost model training service",
    version="1.0.0"
)

# 配置 CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# 创建 Redis 客户端
redis_client = redis.from_url(
    os.getenv("REDIS_URL", "redis://localhost:6379"),
    decode_responses=True
)

# 健康检查
@app.get("/health")
def health_check():
    return {"status": "ok"}

# 训练相关路由
@app.post("/train")
def start_training():
    """启动训练任务"""
    return {
        "status": "success",
        "message": "Training started"
    }

@app.get("/train/{experiment_id}/status")
def get_training_status(experiment_id: str):
    """获取训练状态"""
    return {
        "experiment_id": experiment_id,
        "status": "pending",
        "progress": 0
    }

@app.post("/train/{experiment_id}/stop")
def stop_training(experiment_id: str):
    """停止训练任务"""
    return {
        "status": "success",
        "message": "Training stopped"
    }

# 模型相关路由
@app.get("/models/{model_id}")
def get_model(model_id: str):
    """获取模型信息"""
    return {
        "model_id": model_id,
        "name": "Sample Model",
        "status": "completed"
    }

@app.get("/models/{model_id}/download")
def download_model(model_id: str):
    """下载模型文件"""
    return {
        "status": "success",
        "message": "Model download started"
    }

if __name__ == "__main__":
    uvicorn.run(
        "app.main:app",
        host="0.0.0.0",
        port=int(os.getenv("PORT", "8000")),
        reload=True
    )