apps/trainer/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
xgboost>=2.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
redis>=5.0.0
pydantic>=2.6.0
httpx>=0.27.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0